+++ CHANGELOG.md
@@ -15,6 +15,7 @@ in 8.x, 7.x, 6.x, 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
 ## Lab
 
 ### 2026-02-09
+  * Added Smart Distribute spacing feature
   * Updated PackingSolver dependency
   * Improved Smart Stretch Tool performance
   * Improved management of locked states in the Smart Stretch tool